requirements.txt
Python 3.9+
streamlit
pandas
scikit-learn
joblib
Pillow
